+++ ZucchiniCore/Entities/UserLikedArticle.cs
@@ -5,7 +5,7 @@ public class UserLikedArticle
     public string UserId { get; set; } = string.Empty;
     public int ArticleId { get; set; }
     public User User { get; set; } = null!;
-    public Article Article { get; set; } = null!;
+
 }
