+++ ZucchiniMVC/Application/Services/Recommendation/IRecommendationService.cs
@@ -1,0 +1,8 @@
+﻿using ZucchiniCore.Entities;
+
+namespace ZucchiniMVC.Application.Services.RecommendationServcie;
+
+public interface IRecommendationService
+{
+    Task<List<Article>> GetRecommendArticles(string userId);
+}
